# server/docs/swagger.yaml
definitions:
...
  request.Register:
    properties:
      authorityId:
        maxLength: 20
        type: string
      email:
        type: string
      orgId:
        minimum: 1
        type: integer
      userName:
        maxLength: 30
        type: string
    required:
    - authorityId
    - email
    - orgId
    type: object
...
  response.Response:
    properties:
      code:
        type: integer
      message:
        type: string
      result: {}
      type:
        type: string
    type: object
...
  tms.ManufacturerDTO:
    properties:
      address:
        maxLength: 100
        type: string
      city:
        description: Optional data
        maxLength: 100
        type: string
      contactName:
        maxLength: 50
        minLength: 3
        type: string
      country:
        description: TODO:"binding:"required,iso3166_1_alpha2"
        type: string
      description:
        maxLength: 255
        type: string
      email:
        type: string
      name:
        maxLength: 50
        minLength: 3
        type: string
      phone:
        type: string
      postalCode:
        description: binding:"postcode_iso3166_alpha2=GB"
        type: string
      signatureProvider:
        maxLength: 255
        type: string
    required:
    - contactName
    - country
    - email
    - name
    - phone
    type: object
...
info:
  contact: {}
  description: This is a sample Server pets
  title: Swagger Example API
  version: 0.0.1
paths:
...
  /manufacturer/UpdateManufacturerStatus/{id}:
    put:
      consumes:
      - application/json
      parameters:
      - description: Manufacturer id
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: '{"code": 0,"message": "DataSuccess","result": {},"type": "success"}'
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: '{"code": 7,"message": "Data Fail","result": {},"type": "fail"}'
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - ApiKeyAuth: []
      summary: UpdateManufacturerStatus
      tags:
      - Manufacturer
  /manufacturer/createManufacturer:
    post:
      consumes:
      - application/json
      parameters:
      - description: createManufacturer
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/tms.ManufacturerDTO'
      produces:
      - application/json
      responses:
        "200":
          description: '{"code": 0,"message": "DataSuccess","result": {},"type": "success"}'
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: '{"code": 7,"message": "Bad Request","result": {},"type": "fail"}'
          schema:
            $ref: '#/definitions/response.Response'
        "409":
          description: '{"code": 7,"message": "conflict data","result": {},"type":
            "fail"}'
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: '{"code": 7,"message": "create fail","result": {},"type": "fail"}'
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - ApiKeyAuth: []
      summary: createManufacturer
      tags:
      - Manufacturer
  /manufacturer/deleteManufacturer/{id}:
    delete:
      parameters:
      - description: Manufacturer id
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: '{"code": 0,"message": "delete Success","result": {},"type":
            "success"}'
          schema:
            $ref: '#/definitions/response.Response'
        "404":
          description: '{"code": 7,"message": "idNotFound","result": {},"type": "fail"}'
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: '{"code": 7,"message": "Data Fail","result": {},"type": "fail"}'
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - ApiKeyAuth: []
      summary: deleteManufacturer
      tags:
      - Manufacturer
  /manufacturer/exportManufacturerExcel:
    get:
      consumes:
      - Application/json
      parameters:
      - in: query
        name: country
        type: string
      - in: query
        name: createdAfter
        type: string
      - in: query
        name: createdBefore
        type: string
      - in: query
        name: email
        type: string
      - description: PDF || Excel
        in: query
        name: format
        type: string
      - in: query
        name: name
        type: string
      - description: ASC ||  default= DESC|
        in: query
        name: order
        type: string
      - description: page number
        in: query
        name: page
        type: integer
      - description: page size
        in: query
        name: pageSize
        type: integer
      - in: query
        name: phone
        type: string
      - description: platform=Android|monitor|prolin
        in: query
        name: platform
        type: string
      - in: query
        name: status
        type: string
      produces:
      - Application/json
      responses:
        "200":
          description: '{"code": 0,"message": "DataSuccess","result": {},"type": "success"}'
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal Server Error
      security:
      - ApiKeyAuth: []
      summary: exportManufacturerExcel
      tags:
      - Manufacturer
  /manufacturer/findManufacturer/{id}:
    get:
      parameters:
      - description: queryByIdManufacturer
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: '{"code": 0,"message": "DataSuccess","result": {},"type": "success"}'
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: '{"code": 7,"message": "Bad Request","result": {},"type": "fail"}'
          schema:
            $ref: '#/definitions/response.Response'
        "404":
          description: '{"code": 7,"message": "idNotFound","result": {},"type": "fail"}'
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: '{"code": 7,"message": "Data Fail","result": {},"type": "fail"}'
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - ApiKeyAuth: []
      summary: queryByIdManufacturer
      tags:
      - Manufacturer
  /manufacturer/getManufacturerList:
    get:
      consumes:
      - Application/json
      parameters:
      - in: query
        name: country
        type: string
      - in: query
        name: createdAfter
        type: string
      - in: query
        name: createdBefore
        type: string
      - in: query
        name: email
        type: string
      - description: PDF || Excel
        in: query
        name: format
        type: string
      - in: query
        name: name
        type: string
      - description: ASC ||  default= DESC|
        in: query
        name: order
        type: string
      - description: page number
        in: query
        name: page
        type: integer
      - description: page size
        in: query
        name: pageSize
        type: integer
      - in: query
        name: phone
        type: string
      - description: platform=Android|monitor|prolin
        in: query
        name: platform
        type: string
      - in: query
        name: status
        type: string
      produces:
      - Application/json
      responses:
        "200":
          description: '{"code": 0,"message": "DataSuccess","result": {},"type": "success"}'
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal Server Error
      security:
      - ApiKeyAuth: []
      summary: GetManufacturerList
      tags:
      - Manufacturer
  /manufacturer/updateManufacturer/{id}:
    put:
      consumes:
      - application/json
      parameters:
      - description: updateManufacturer
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/tms.ManufacturerDTO'
      - description: Manufacturer id
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: '{"code": 0,"message": "DataSuccess","result": {},"type": "success"}'
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - ApiKeyAuth: []
      summary: updateManufacturer
      tags:
      - Manufacturer
...
  /user/register:
    post:
      parameters:
      - description: username, password, roleID, orgID
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/request.Register'
      produces:
      - application/json
      responses:
        "200":
          description: User registered account,return includes user info
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.Response'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Response'
      summary: User registered account
      tags:
      - SysUser
...
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"